bin/clean/requirements.txt
pyyaml>=3.11
requests>=2.25
